move fromSecureRandom to UUIDGen
