+++ std/js-native/src/main/scala/cats/effect/std/UUIDGenCompanionPlatform.scala
@@ -32,21 +32,34 @@ import cats.effect.kernel.Sync
 
 import java.util.UUID
 
-private[std] trait UUIDGenCompanionPlatform {
-  implicit def fromSync[F[_]](implicit ev: Sync[F]): UUIDGen[F] = new UUIDGen[F] {
-    private val csprng = new SecureRandom.JavaSecureRandom()
-    private val randomUUIDBuffer = new Array[Byte](16)
-    override final val randomUUID: F[UUID] =
-      ev.delay {
-        val buffer = randomUUIDBuffer // local copy
+private[std] trait UUIDGenCompanionPlatform extends UUIDGenCompanionPlatformLowPriority
 
-        /* We use nextBytes() because that is the primitive of most secure RNGs,
-         * and therefore it allows to perform a unique call to the underlying
-         * secure RNG.
-         */
-        csprng.nextBytes(randomUUIDBuffer)
+private[std] trait UUIDGenCompanionPlatformLowPriority {
+  implicit def fromSync[F[_]](implicit ev: Sync[F]): UUIDGen[F] =
+    new UUIDGen[F] {
+      private val csprng = new SecureRandom.JavaSecureRandom()
+      private val randomUUIDBuffer = new Array[Byte](16)
+      override final val randomUUID: F[UUID] =
+        ev.delay {
+          val buffer = randomUUIDBuffer // local copy
 
-        UnsafeUUIDBuilder.build(buffer)
-      }
-  }
+          /* We use nextBytes() because that is the primitive of most secure RNGs,
+           * and therefore it allows to perform a unique call to the underlying
+           * secure RNG.
+           */
+          csprng.nextBytes(randomUUIDBuffer)
+
+          @inline def intFromBuffer(i: Int): Int =
+            (buffer(i) << 24) | ((buffer(i + 1) & 0xff) << 16) | ((buffer(
+              i + 2) & 0xff) << 8) | (buffer(i + 3) & 0xff)
+
+          val i1 = intFromBuffer(0)
+          val i2 = (intFromBuffer(4) & ~0x0000f000) | 0x00004000
+          val i3 = (intFromBuffer(8) & ~0xc0000000) | 0x80000000
+          val i4 = intFromBuffer(12)
+          val msb = (i1.toLong << 32) | (i2.toLong & 0xffffffffL)
+          val lsb = (i3.toLong << 32) | (i4.toLong & 0xffffffffL)
+          new UUID(msb, lsb)
+        }
+    }
 }
