+++ std/js/src/main/scala/cats/effect/std/SecureRandomCompanionPlatform.scala
@@ -28,6 +28,11 @@
 
 package cats.effect.std
 
+import cats.Applicative
+import cats.effect.kernel.Sync
+import cats.effect.std.Random.ScalaRandom
+import cats.syntax.all._
+
 import scala.scalajs.js
 import scala.scalajs.js.typedarray._
 
@@ -109,5 +114,10 @@ private[std] trait SecureRandomCompanionPlatform {
       )
     }
   }
+
+  def javaSecuritySecureRandom[F[_]: Sync]: F[SecureRandom[F]] =
+    Sync[F]
+      .delay(new JavaSecureRandom())
+      .map(r => new ScalaRandom[F](Applicative[F].pure(r)) with SecureRandom[F] {})
 
 }
