+++ std/jvm/src/main/scala/cats/effect/std/UUIDGenCompanionPlatform.scala
@@ -16,20 +16,11 @@
 
 package cats.effect.std
 
-import cats.Functor
 import cats.effect.kernel.Sync
-import cats.syntax.all._
 
 import java.util.UUID
 
-private[std] trait UUIDGenCompanionPlatform extends UUIDGenCompanionPlatformLowPriority {
-  implicit def fromSecureRandom[F[_]: Functor: SecureRandom]: UUIDGen[F] = {
-    new UUIDGen[F] {
-      override final val randomUUID: F[UUID] =
-        SecureRandom[F].nextBytes(16).map(UnsafeUUIDBuilder.build)
-    }
-  }
-}
+private[std] trait UUIDGenCompanionPlatform extends UUIDGenCompanionPlatformLowPriority
 
 private[std] trait UUIDGenCompanionPlatformLowPriority {
